+++ packages/react-form/src/index.ts
@@ -22,4 +22,5 @@ export {
   propagateErrors,
   validateAll,
   reduceFields,
+  makeCleanFields,
 } from './utilities';